fixed user creation api

+++ Backend/src/main/java/com/ebdapo/backend/restcontroller/ApothekenController.java
@@ -27,6 +27,7 @@ public class ApothekenController {
 
     @PostMapping("/apotheke")
     public ResponseEntity<?> newApotheke(@RequestBody Apotheke apotheke) {
+
         if(apotheke.getName() == null || apotheke.getEmail() == null || apotheke.getAnschrift() == null) {
             throw new InvalidInputException("Ungültige oder fehlende Angaben");
         }
@@ -53,7 +54,8 @@ public class ApothekenController {
 
     @PutMapping("/apotheke/{apothekeId}")
     public ResponseEntity<?> updateApotheke(@PathVariable String apothekeId, @RequestBody Apotheke newApo) {
-        if(!authController.checkIfAuthorized(authController.getCurrentUsername(), apothekeId)) {
+        //only admin can update apotheke
+        if(!authController.isAdmin(authController.getCurrentUsername(), apothekeId)) {
             return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
 
@@ -74,9 +76,15 @@ public class ApothekenController {
 
     @DeleteMapping("/apotheke/{apothekeId}")
     public ResponseEntity<?> deleteApotheke(@PathVariable String apothekeId) {
+        //only admin can delete apotheke
+        if(!authController.isAdmin(authController.getCurrentUsername(), apothekeId)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         if(!apothekenRepo.existsById(apothekeId)) {
             throw new BadRequestException();
         }
+
         apothekenRepo.deleteById(apothekeId);
         return new ResponseEntity<>(HttpStatus.OK);
     }
